refactor(auth): remove dead nostrify login stack; split session context/provider (#178)
useCurrentUser no longer reads the nostrify login store, leaving LoginArea,
LoginDialog, SignupDialog, AccountSwitcher, useLoginActions, useLoggedInAccounts
and NostrLoginProvider as fully bypassed dead code (two parallel identity
systems). Remove them. Also split DivineSessionContext into a context-only
module + a DivineSessionProvider component, matching the repo's AppContext/
AppProvider pattern and clearing the react-refresh lint warning.
