--- src/hooks/useLoginActions.ts
@@ -1,34 +1,0 @@
-import { useNostr } from '@nostrify/react';
-import { NLogin, useNostrLogin } from '@nostrify/react/login';
-
-// NOTE: This file should not be edited except for adding new login methods.
-
-export function useLoginActions() {
-  const { nostr } = useNostr();
-  const { logins, addLogin, removeLogin } = useNostrLogin();
-
-  return {
-    // Login with a Nostr secret key
-    nsec(nsec: string): void {
-      const login = NLogin.fromNsec(nsec);
-      addLogin(login);
-    },
-    // Login with a NIP-46 "bunker://" URI
-    async bunker(uri: string): Promise<void> {
-      const login = await NLogin.fromBunker(uri, nostr);
-      addLogin(login);
-    },
-    // Login with a NIP-07 browser extension
-    async extension(): Promise<void> {
-      const login = await NLogin.fromExtension();
-      addLogin(login);
-    },
-    // Log out the current user
-    async logout(): Promise<void> {
-      const login = logins[0];
-      if (login) {
-        removeLogin(login.id);
-      }
-    }
-  };
-}
